+++ lib/src/model/page_view_model.dart
@@ -5,8 +5,14 @@ class PageViewModel {
   /// Title of page
   final String title;
 
+  /// Title of page
+  final Widget titleWidget;
+
   /// Text of page (description)
   final String body;
+
+  /// Widget content of page (description)
+  final Widget bodyWidget;
 
   /// Image of page
   /// Tips: Wrap your image with an alignment widget like Align or Center
@@ -19,12 +25,28 @@ class PageViewModel {
   /// Contain all page customizations, like page color, text styles
   final PageDecoration decoration;
 
-  PageViewModel(
+  PageViewModel({
     this.title,
-    this.body, {
+    this.titleWidget,
+    this.body,
+    this.bodyWidget,
     this.image,
     this.footer,
     this.decoration = const PageDecoration(),
-  })  : assert(title != null, "You can not provide a null title text"),
-        assert(body != null, "You can not provide a null body text");
+  })  : assert(
+          title != null || titleWidget != null,
+          "You must provide either title (String) or titleWidget (Widget).",
+        ),
+        assert(
+          (title == null) != (titleWidget == null),
+          "You can not provide both title and titleWidget.",
+        ),
+        assert(
+          body != null || bodyWidget != null,
+          "You must provide either body (String) or bodyWidget (Widget).",
+        ),
+        assert(
+          (body == null) != (bodyWidget == null),
+          "You can not provide both body and bodyWidget.",
+        );
 }